requirements.txt
   
aiogram
aiohttp
aiofiles
python-dotenv